use Trait to reduce code duplication for aggregation functions
