+++ src/Queries/Mysql/Count.php
@@ -4,14 +4,14 @@ namespace SimpleCrud\Queries\Mysql;
 
 use SimpleCrud\Queries\Query;
 use SimpleCrud\Table;
-use PDO;
 
 /**
  * Manages a database select count query.
  */
 class Count extends Query
 {
-    use ExtendedSelectionTrait;
+    use AggregationTrait;
+    const AGGREGATION_FUNCTION = 'COUNT';
 
     /**
      * Returns the count.
@@ -27,23 +27,13 @@ class Count extends Query
         return (int) $result[0];
     }
 
-    /**
-     * {@inheritdoc}
-     */
-    public function __invoke()
-    {
-        $statement = $this->table->getDatabase()->execute((string) $this, $this->marks);
-        $statement->setFetchMode(PDO::FETCH_NUM);
-
-        return $statement;
-    }
 
     /**
      * {@inheritdoc}
      */
     public function __toString()
     {
-        $query = "SELECT COUNT(*) FROM `{$this->table->getName()}`";
+        $query = "SELECT ".self::AGGREGATION_FUNCTION."(*) FROM `{$this->table->getName()}`";
 
         $query .= $this->fromToString();
         $query .= $this->whereToString();
