+++ src/Queries/Mysql/Min.php
@@ -4,68 +4,13 @@ namespace SimpleCrud\Queries\Mysql;
 
 use SimpleCrud\Queries\Query;
 use SimpleCrud\Table;
-use PDO;
+
 
 /**
  * Manages a database select min query in Mysql databases.
  */
 class Min extends Query
 {
-    use ExtendedSelectionTrait;
-
-    protected $field;
-
-    /**
-     * Set the field name to maximize over.
-     *
-     * @param string $field
-     *
-     * @return self
-     */
-    public function field($field)
-    {
-        $this->field = $field;
-
-        return $this;
-    }
-
-    /**
-     * Run the query and return the value.
-     * 
-     * @return int
-     */
-    public function run()
-    {
-        $result = $this->__invoke()->fetch();
-        $field = $this->table->{$this->field};
-        
-        return $field->dataFromDatabase($result[0]);
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function __invoke()
-    {
-        $statement = $this->table->getDatabase()->execute((string) $this, $this->marks);
-        $statement->setFetchMode(PDO::FETCH_NUM);
-
-        return $statement;
-    }
-
-    /**
-     * Build and return the query.
-     *
-     * @return string
-     */
-    public function __toString()
-    {
-        $query = "SELECT MIN(`{$this->field}`) FROM `{$this->table->getName()}`";
-
-        $query .= $this->fromToString();
-        $query .= $this->whereToString();
-        $query .= $this->limitToString();
-
-        return $query;
-    }
+    use AggregationTrait;
+    const AGGREGATION_FUNCTION = 'MIN';
 }
